Add tests for LatestVersionFinder
